permissions for feedbacks and feedback responses

+++ packages/graphql/src/services/feedbacks.ts
@@ -100,7 +100,16 @@ export async function respondToFeedback(
   { id, responseContent }: { id: number; responseContent: string },
   ctx: ContextWithUser
 ) {
-  const feedback = await ctx.prisma.feedback.update({
+  const feedback = await ctx.prisma.feedback.findUnique({
+    where: { id },
+    include: {
+      session: true,
+    },
+  })
+
+  if (!feedback || feedback.session.ownerId !== ctx.user.sub) return null
+
+  const updatedFeedback = await ctx.prisma.feedback.update({
     where: { id },
     data: {
       isPublished: true,
@@ -118,14 +127,14 @@ export async function respondToFeedback(
     },
   })
 
-  ctx.pubSub.publish('feedbackUpdated', feedback)
+  ctx.pubSub.publish('feedbackUpdated', updatedFeedback)
 
   ctx.emitter.emit('invalidate', {
     typename: 'Session',
-    id: feedback.sessionId,
+    id: updatedFeedback.sessionId,
   })
 
-  return feedback
+  return updatedFeedback
 }
 
 // add confusion timestep to session
@@ -163,7 +172,16 @@ export async function publishFeedback(
   { id, isPublished }: { id: number; isPublished: boolean },
   ctx: ContextWithUser
 ) {
-  const feedback = await ctx.prisma.feedback.update({
+  const feedback = await ctx.prisma.feedback.findUnique({
+    where: { id },
+    include: {
+      session: true,
+    },
+  })
+
+  if (!feedback || feedback.session.ownerId !== ctx.user.sub) return null
+
+  const updatedFeedback = await ctx.prisma.feedback.update({
     where: {
       id,
     },
@@ -176,17 +194,17 @@ export async function publishFeedback(
   })
 
   if (isPublished) {
-    ctx.pubSub.publish('feedbackAdded', feedback)
+    ctx.pubSub.publish('feedbackAdded', updatedFeedback)
   } else {
-    ctx.pubSub.publish('feedbackRemoved', feedback)
+    ctx.pubSub.publish('feedbackRemoved', updatedFeedback)
   }
 
   ctx.emitter.emit('invalidate', {
     typename: 'Session',
-    id: feedback.sessionId,
+    id: updatedFeedback.sessionId,
   })
 
-  return feedback
+  return updatedFeedback
 }
 
 // pin / unpin a feedback on the lecturers running session screen
@@ -194,7 +212,16 @@ export async function pinFeedback(
   { id, isPinned }: { id: number; isPinned: boolean },
   ctx: ContextWithUser
 ) {
-  const feedback = await ctx.prisma.feedback.update({
+  const feedback = await ctx.prisma.feedback.findUnique({
+    where: { id },
+    include: {
+      session: true,
+    },
+  })
+
+  if (!feedback || feedback.session.ownerId !== ctx.user.sub) return null
+
+  const updatedFeedback = await ctx.prisma.feedback.update({
     where: {
       id,
     },
@@ -206,14 +233,14 @@ export async function pinFeedback(
     },
   })
 
-  ctx.pubSub.publish('feedbackUpdated', feedback)
+  ctx.pubSub.publish('feedbackUpdated', updatedFeedback)
 
   ctx.emitter.emit('invalidate', {
     typename: 'Session',
-    id: feedback.sessionId,
+    id: updatedFeedback.sessionId,
   })
 
-  return feedback
+  return updatedFeedback
 }
 
 // resolve / unresolve a feedback
@@ -221,7 +248,16 @@ export async function resolveFeedback(
   { id, isResolved }: { id: number; isResolved: boolean },
   ctx: ContextWithUser
 ) {
-  const feedback = await ctx.prisma.feedback.update({
+  const feedback = await ctx.prisma.feedback.findUnique({
+    where: { id },
+    include: {
+      session: true,
+    },
+  })
+
+  if (!feedback || feedback.session.ownerId !== ctx.user.sub) return null
+
+  const updatedFeedback = await ctx.prisma.feedback.update({
     where: { id },
     data: {
       isResolved: isResolved,
@@ -232,14 +268,14 @@ export async function resolveFeedback(
     },
   })
 
-  ctx.pubSub.publish('feedbackUpdated', feedback)
+  ctx.pubSub.publish('feedbackUpdated', updatedFeedback)
 
   ctx.emitter.emit('invalidate', {
     typename: 'Session',
-    id: feedback.sessionId,
+    id: updatedFeedback.sessionId,
   })
 
-  return feedback
+  return updatedFeedback
 }
 
 // deletes a feedback (and all its responses through cascade)
@@ -247,18 +283,30 @@ export async function deleteFeedback(
   { id }: { id: number },
   ctx: ContextWithUser
 ) {
-  const feedback = await ctx.prisma.feedback.delete({
+  const feedback = await ctx.prisma.feedback.findUnique({
+    where: { id },
+    include: {
+      session: true,
+    },
+  })
+
+  if (!feedback || feedback.session.ownerId !== ctx.user.sub) return null
+
+  const deletedFeedback = await ctx.prisma.feedback.delete({
     where: { id },
   })
 
-  ctx.pubSub.publish('feedbackRemoved', { id, sessionId: feedback.sessionId })
+  ctx.pubSub.publish('feedbackRemoved', {
+    id,
+    sessionId: deletedFeedback.sessionId,
+  })
 
   ctx.emitter.emit('invalidate', {
     typename: 'Session',
-    id: feedback.sessionId,
+    id: deletedFeedback.sessionId,
   })
 
-  return feedback
+  return deletedFeedback
 }
 
 // deletes a feedback response
@@ -266,7 +314,24 @@ export async function deleteFeedbackResponse(
   { id }: { id: number },
   ctx: ContextWithUser
 ) {
-  const feedbackResponse = await ctx.prisma.feedbackResponse.delete({
+  const feedbackResponse = await ctx.prisma.feedbackResponse.findUnique({
+    where: { id },
+    include: {
+      feedback: {
+        include: {
+          session: true,
+        },
+      },
+    },
+  })
+
+  if (
+    !feedbackResponse ||
+    feedbackResponse.feedback.session.ownerId !== ctx.user.sub
+  )
+    return null
+
+  const deletedFeedbackResponse = await ctx.prisma.feedbackResponse.delete({
     where: { id },
     include: {
       feedback: {
@@ -277,12 +342,12 @@ export async function deleteFeedbackResponse(
     },
   })
 
-  ctx.pubSub.publish('feedbackUpdated', feedbackResponse.feedback)
+  ctx.pubSub.publish('feedbackUpdated', deletedFeedbackResponse.feedback)
 
   ctx.emitter.emit('invalidate', {
     typename: 'Session',
-    id: feedbackResponse.feedback.sessionId,
+    id: deletedFeedbackResponse.feedback.sessionId,
   })
 
-  return feedbackResponse.feedback
+  return deletedFeedbackResponse.feedback
 }
